+++ src/scheduler.py
@@ -3,6 +3,7 @@
 import asyncio
 from collections import defaultdict
 import configparser
+from datetime import datetime
 from functools import partial
 from json import dumps as json_dumps, loads as json_loads
 from re import match as re_match
@@ -14,12 +15,13 @@ from unittest.mock import Mock
 
 from bidict import bidict
 from kubernetes import client, config, watch
+import pytz
 
 from health_server import HealthServer
 
 
 class Scheduler():
-    def __init__(self, name, reserved_kublet_cpu, reserved_kublet_ram, optimizer_port):
+    def __init__(self, name, reserved_kublet_cpu, reserved_kublet_ram, no_solution_found_warning, optimizer_port):
         '''Load kubernetes config and connect to the API server.'''
         try:
             config.load_incluster_config()
@@ -29,6 +31,7 @@ class Scheduler():
         self.name = name
         self.reserved_kublet_cpu = reserved_kublet_cpu
         self.reserved_kublet_ram = reserved_kublet_ram
+        self.no_solution_found_warning = no_solution_found_warning
         self.optimizer = self.Optimizer(optimizer_port)
         self.nicknames = bidict() # we naively assume names don't change during runtime; TODO don't
 
@@ -240,6 +243,20 @@ class Scheduler():
         self.api.create_namespaced_binding(namespace=namespace, body=body, _preload_content=False)
 
 
+    def warn_no_solution_found(self, name, namespace='default'):
+        '''Add event message to pod description when optimizer is unable to find solution.'''
+        meta = client.V1ObjectMeta(name = name)
+        reference = client.V1ObjectReference(name = name, kind = 'Pod')
+        body = client.V1Event(metadata = meta,
+                              involved_object = reference,
+                              last_timestamp = datetime.now(pytz.utc),
+                              type = 'Warning',
+                              reason = 'FailedScheduling',
+                              reporting_component = self.name,
+                              message = "Optimizer was unable to find solution for batch containing pod.")
+        self.api.create_namespaced_event(namespace=namespace, body=body, _preload_content=False)
+
+
     def wait_for_optimizer(self):
         '''Wait for optimizer's readiness probe to return HTTP 200.'''
         response = Mock(spec=Response)
@@ -297,6 +314,9 @@ class Scheduler():
                 raise Exception(json_loads(e.body)['message'])
         else:
             print('Warning: no configuration returned from optimizer: {}'.format(result['error']))
+            if self.no_solution_found_warning:
+                for event in regular_events:
+                    self.warn_no_solution_found(event['object'].metadata.name)
 
 
     def get_event_batch(self, previous, batch_limit, time_limit):
@@ -336,6 +356,7 @@ if __name__ == '__main__':
     scheduler = Scheduler(settings['Boreas']['SchedulerName'],
                           reserved_kublet_cpu=settings.getint('Boreas', 'ReservedKubletCPU'),
                           reserved_kublet_ram=settings.getint('Boreas', 'ReservedKubletRAM'),
+                          no_solution_found_warning=settings.getboolean('Boreas', 'WarnNoSolutionFound'),
                           optimizer_port=settings.getint('Boreas', 'OptimizerPort'))
     asyncio.run(scheduler.start(batch_limit=settings.getint('Boreas', 'BatchSize'),
                                 time_limit=settings.getint('Boreas', 'BatchTime')))